starting hirise tests
