+++ isis/src/mro/apps/hicubenorm/hicubenorm.h
@@ -1,0 +1,12 @@
+#ifndef hicubenorm_h 
+#define hicubenorm_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis {
+  extern void hicubenorm(Cube *cube, UserInterface &ui);
+  extern void hicubenorm(UserInterface &ui);
+}
+
+#endif
